Continued Phase 1 fixes: CS8121 pattern matching, CS0173 conditional types, CS0019 operator errors, CS1501 method overloads - down to ~10 critical CS errors

+++ src/BotCore/Bandits/NeuralUcbExtended.cs
@@ -459,7 +459,7 @@ public class NeuralUcbExtended : IDisposable
             
         // Volume volatility indicator
         var volumeVolatility = marketContext.Volume > 0 
-            ? (decimal)Math.Min(marketContext.Volume / 1000000.0, 1.0) // Normalize volume
+            ? (decimal)Math.Min((double)marketContext.Volume / 1000000.0, 1.0) // Normalize volume
             : 0m;
         
         // Weighted combination favoring spread and ATR
